requirements.txt
ansible==2.5.2
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
confusable-homoglyphs==3.0.0
coverage==4.5.1
coveralls==1.3.0
cryptography==2.2.2
Django==2.0.4
django-appconf==1.0.2
django-multiselectfield==0.1.8
django-registration==2.4.1
django-sass-processor==0.7
djangorestframework==3.8.2
docopt==0.6.2
factory-boy==2.10.0
Faker==0.8.13
flake8==3.5.0
idna==2.6
Jinja2==2.10
libsass==0.14.5
MarkupSafe==1.0
mccabe==0.6.1
paramiko==2.4.1
pep8==1.7.1
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.4
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyNaCl==1.2.1
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
urllib3==1.22
